Adjustments to new project structure
