+++ hc_core/Cargo.toml
@@ -4,6 +4,8 @@ version = "0.1.0"
 authors = ["Nicolas Luck <nicolas@lucksus.org>"]
 
 [dependencies]
+wasmi = { git = "https://github.com/paritytech/wasmi.git"}
+wabt = "0.2.0"
 
 [features]
 # Treat warnings as a build error.
